backend/requirements.txt
fastapi
python-jose[cryptography]
passlib[bcrypt]
uvicorn
httpx
transformers
accelerate
soundfile
sqlalchemy
psycopg2-binary